+++ server/src/au/com/codeka/warworlds/server/handlers/RankingHistoryHandler.java
@@ -41,11 +41,11 @@ public class RankingHistoryHandler extends RequestHandler {
                 Messages.EmpireRank.Builder empire_rank_pb = Messages.EmpireRank.newBuilder();
                 empire_rank_pb.setEmpireKey(Integer.toString(res.getInt(1)));
                 empire_rank_pb.setRank(rank);
-                empire_rank_pb.setTotalStars(BigInteger.valueOf(res.getLong(3)));
-                empire_rank_pb.setTotalColonies(BigInteger.valueOf(res.getLong(4)));
-                empire_rank_pb.setTotalBuildings(BigInteger.valueOf(res.getLong(5)));
-                empire_rank_pb.setTotalShips(BigInteger.valueOf(res.getLong(6)));
-                empire_rank_pb.setTotalPopulation(BigInteger.valueOf(res.getLong(7)));
+                empire_rank_pb.setTotalStars(res.getLong(3));
+                empire_rank_pb.setTotalColonies(res.getLong(4));
+                empire_rank_pb.setTotalBuildings(res.getLong(5));
+                empire_rank_pb.setTotalShips(res.getLong(6));
+                empire_rank_pb.setTotalPopulation(res.getLong(7));
                 empire_ranks_pb.addRanks(empire_rank_pb);
             }
