+++ server/src/au/com/codeka/warworlds/server/model/EmpireRank.java
@@ -1,7 +1,6 @@
 package au.com.codeka.warworlds.server.model;
 
 import java.sql.SQLException;
-import java.math.BigInteger;
 
 import au.com.codeka.common.model.BaseEmpireRank;
 import au.com.codeka.warworlds.server.data.SqlResult;
@@ -17,11 +16,11 @@ public class EmpireRank extends BaseEmpireRank {
 
         try {
             mRank = res.getInt("rank");
-            mTotalStars = BigInteger.valueOf(res.getLong("total_stars"));
-            mTotalColonies = BigInteger.valueOf(res.getLong("total_colonies"));
-            mTotalBuildings = BigInteger.valueOf(res.getLong("total_buildings"));
-            mTotalShips = BigInteger.valueOf(res.getLong("total_ships"));
-            mTotalPopulation = BigInteger.valueOf(res.getLong("total_population"));
+            mTotalStars = res.getLong("total_stars");
+            mTotalColonies = res.getLong("total_colonies");
+            mTotalBuildings = res.getLong("total_buildings");
+            mTotalShips = res.getLong("total_ships");
+            mTotalPopulation = res.getLong("total_population");
         } catch (SQLException e) {
             // these may not exist... doesn't matter
         }
@@ -31,19 +30,19 @@ public class EmpireRank extends BaseEmpireRank {
         return mEmpireID;
     }
 
-    public void setTotalShips(BigInteger num) {
+    public void setTotalShips(long num) {
         mTotalShips = num;
     }
-    public void setTotalBuildings(BigInteger num) {
+    public void setTotalBuildings(long num) {
         mTotalBuildings = num;
     }
-    public void setTotalColonies(BigInteger num) {
+    public void setTotalColonies(long num) {
         mTotalColonies = num;
     }
-    public void setTotalStars(BigInteger num) {
+    public void setTotalStars(long num) {
         mTotalStars = num;
     }
-    public void setTotalPopulation(BigInteger num) {
+    public void setTotalPopulation(long num) {
         mTotalPopulation = num;
     }
 }
